Added Gitter sidecar to Topsoil
Fixes #60 by adding a Gitter sidecar to the Topsoil project page.
Upon navigating to the Topsoil project page, the user now sees an
"Open Chat" tab in the lower right corner. Clicking it opens the
Gitter sidecar, and the user can participate in the CIRDLES/Topsoil
Gitter chat room.

In the process, I added a new include, gitter_sidecar.html. It can add
the "Open Chat" tab for any room to any page as follows:

{% include gitter_sidecar.html room="ROOM_NAME_HERE" %}